feat: complete cli, modules
