+++ src/cli/env.js
@@ -1,5 +1,10 @@
+
 const parseEnv = () => {
-    // Write your code here 
+  const rssEnv = Object.entries(process.env)
+    .filter(([key]) => key.startsWith('RSS_'))
+    .map(([key, value]) => `${key}=${value}`);
+
+  console.log(rssEnv.join('; '));
 };
 
 parseEnv();
